refactor: 객체의 역할에 맞는 테스트 분리
